Optimize execution on FVP and Foundation platform
Add an extra partition to the disk image generated for FVP and
Foundation platform (to be used for user data or extra usage). The
partition is created on the hard drive but is not formated.
Speed up running on FVP and Foundation by optimizing model parameters:
- disable options that are already set by default
- disable cache modelling on FVP
- disable secure memory
- disable real time on Foundation (prevents timeout as execution is a
lot slower then on a real system)

Change-Id: I20d613bd8b0f4cb7df87c294d45b9f718f6a8ed3
Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>