arm/hafnium: refuse to build on non-x86-64 hosts
For reasons, hafnium doesn't use host binaries for stuff like GN, Ninja,
Clang, GCC, or Qemu. Instead it ships binaries that it uses during the
build, but these binaries are only available for x86-64.

Attempts to use external tools (so the recipe can just DEPENDS as usual)
were not successful (clang being the problem) so just forcible stop the
recipe building on anything but x86-64 for now.

Change-Id: I144309d15fa1265b2b105981cd2bfcd376aab7ad
Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>