arm-autonomy/sysvinit-inittab: Fix console login for host and guest
In meta-virtualization master branch the sysvinit-inittab_virtualization.inc
was introduced to handle the installation of the getty-wrapper for spawing the
console login for /dev/hvc0.

Since we don't make usage of the virtualization DISTRO_FEATURES we have to
manually include the mentioned file when building with arm-autonomy-host and
arm-autonomy-guest DISTRO_FEATURES.

Also, since arm-autonomy-guest doesn't include the xen DISTRO_FEATURES we have
to manually include sysvinit-inittab_xen.inc to append the getty-wrapper entry
to /etc/inittab.

Issue-Id: SCM-2329
Signed-off-by: Diego Sueiro <diego.sueiro@arm.com>
Change-Id: I0f92565de8a9053c0541928c30b8f979757bfbca
Signed-off-by: Jon Mason <jon.mason@arm.com>