arm/trusted-firmware-a: fix incorrect license checksum
TF-A 2.5 has a tweaked license.rst file but the checksum wasn't updated.
The CI didn't catch this because of an attempt at making warnings fatal,
which has the side effect of masking some warnings.

Update the checksum for that file, and add MIT explicitly to the license
list as all of the embedded projects are used under the MIT.

Change-Id: Id39b4c49c0efae30c6452e77b1cdf56e43b792d4
Signed-off-by: Ross Burton <ross.burton@arm.com>