arm-bsp/fvp-base: enable user-mode networking in runfvp
Enable user-mode networking with virtio out of the box.  This way port
forwarding can be configured easily without needing to setup host tap
ports.

Moving forward this may be controlled more by runfvp with the machine
configuration simply saying what configuration parameter to use (in this
case, bp.virtio_net), but this is a good starting point to iterate from.

Change-Id: I0b610df87c7dd2c8e3e213daaa967618babbdc7c
Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>