CI: Add zephyr testimage
Add the ability to run testimage against zephyr machines.
Unfortunately, this exposes a bug in meta-zephyr with TESTIMAGE_AUTO not
working correctly.  So, work around this until it can be fixed upstream.
Also, qemu-cortex-a53 does not successfully pass any of the tests it
successfully builds.  So, don't run testimage until that can be
addressed.

Signed-off-by: Jon Mason <jon.mason@arm.com>