arm/qemuarm64-sbsa: replace with qemu-generic-arm64
Migrate the qemuarm64-sbsa machine to use the generic-arm64 machine as a
base.  This new qemu-generic-arm64 should contain only the parts
necessary to boot the generic-arm64 in qemu (using the SBSA machine).
This allows for a single generic image with testing for SBSA compliance.

NOTE: a unique WIC file is needed due to the inability to pass kernel
bootargs (due to needing DHCP for testimage).

Signed-off-by: Jon Mason <jon.mason@arm.com>