CI: disable use of KVM
oe-core c2a2d47 changed the default of QEMU_USE_KVM to 1, so qemu will
want to use KVM and will error if it can't.  Our CI runners don't have
KVM, so we need to disable this.

In the long term this should be more intelligent as some workers have
KVM and some don't, but this will get successful builds again.

Signed-off-by: Ross Burton <ross.burton@arm.com>