CI: add patch status metrics
Embed an improved patchreview tool which can generate metrics.txt files,
and run that as part of the CI.  This means that every merge request
will include a section if the metrics change, so it is easy to spot if
patches with bad headers are added.

The changes to patchreview will merge into oe-core soon, so when that
happens we can drop the copy.

Signed-off-by: Ross Burton <ross.burton@arm.com>