CI: fix check-layer indeterminism
yocto-check-layers is very greedy when searching for layers, and will
find the test layers in Bitbake if given the chance, for example:

bitbake/lib/layerindexlib/tests/testdata/layer4

This layer has the collection name openembedded-layer and is only
compatible with Sumo.  The selection of layer from collection name is
not deterministic, so it's possible that this layer is selected which
then fails the check as it isn't compatible.

Solve this by restricting the dependency layers in meta-arm to just
meta-arm/meta-*, so it doesn't recurse into meta-arm/work/poky/bitbake.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>