arm-bsp/fvp-base-arm32: use correct DeviceTree
The 64-bit fvp-base machine uses the upstream fvp-base-recv devicetree,
but fvp-base-arm32 was accidentally using the old
fvp-base-gicv3-psci-custom that we patch into the kernel.

As the only difference between these platforms at a "hardware" level is
whether the cores boot in 32- or 64-bit mode, they should both use
fvp-base-revc.

This isn't trivial as devicetree files need to  be under the correct
arch/ directory, so we need to symlink into arch/arm the right files
from arch/arm64.

This has several improvements, but primarily virtio networking works so
we can now use testimage with fvp-base-arm32.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>